@@ -17,38 +17,35 @@
         }
     ],
     "require": {
-        "php": "^7.4|^8.0",
+        "php": "^8.1",
         "ext-iconv": "*",
         "ext-json": "*",
         "ext-mbstring": "*",
         "ext-tokenizer": "*",
-        "cmgmyr/phploc": "^8.0.3",
-        "composer/semver": "^3.4",
-        "friendsofphp/php-cs-fixer": "^3.40.0",
-        "justinrainbow/json-schema": "^5.2.13",
-        "league/container": "^3.2|^4.2",
-        "php-parallel-lint/php-parallel-lint": "^1.3.2",
-        "psr/container": "^1.0|^2.0.2",
-        "psr/simple-cache": "^1.0|^2.0|^3.0",
-        "sebastian/diff": "^4.0|^5.0.3|^6.0",
-        "slevomat/coding-standard": "^8.14.1",
-        "squizlabs/php_codesniffer": "^3.7.2",
-        "symfony/cache": "^5.4|^6.0|^7.0",
-        "symfony/console": "^5.4|^6.4|^7.0",
-        "symfony/finder": "^5.4|^6.0|^7.0",
-        "symfony/http-client": "^5.4|^6.0|^7.0",
-        "symfony/process": "^5.4|^6.4|^7.0"
+        "cmgmyr/phploc": "^8.0.4",
+        "composer/semver": "^3.4.3",
+        "friendsofphp/php-cs-fixer": "^3.74.0",
+        "justinrainbow/json-schema": "^5.3.0",
+        "league/container": "^5.0.1",
+        "php-parallel-lint/php-parallel-lint": "^1.4.0",
+        "psr/container": "^2.0.2",
+        "psr/simple-cache": "^2.0|^3.0",
+        "sebastian/diff": "^5.1.1|^6.0.2",
+        "slevomat/coding-standard": "^8.16.2",
+        "squizlabs/php_codesniffer": "^3.12.0",
+        "symfony/cache": "^6.4.20|^7.2.5",
+        "symfony/console": "^6.4.20|^7.2.5",
+        "symfony/finder": "^6.4.17|^7.2.2",
+        "symfony/http-client": "^6.4.19|^7.2.4",
+        "symfony/process": "^6.4.20|^7.2.5"
     },
     "require-dev": {
-        "ergebnis/phpstan-rules": "^0.15.3",
-        "illuminate/console": "^5.8|^6.0|^7.0|^8.0|^9.20|^10.0",
-        "illuminate/support": "^5.8|^6.0|^7.0|^8.0|^9.52.16|^10.0",
-        "mockery/mockery": "^1.6.6",
-        "phpstan/phpstan-strict-rules": "^0.12.11",
-        "phpunit/phpunit": "^8.0|^9.0|^10.4.2",
-        "rector/rector": "0.11.56",
-        "symfony/var-dumper": "^5.4|^6.0|^7.0",
-        "thecodingmachine/phpstan-strict-rules": "^0.12.2"
+        "illuminate/console": "^10.48.28|^11.44.2|^12.3",
+        "illuminate/support": "^10.48.28|^11.44.2|^12.3",
+        "mockery/mockery": "^1.6.12",
+        "phpstan/phpstan": "^2.1.11",
+        "phpunit/phpunit": "^10.5.45|^11.5.15",
+        "symfony/var-dumper": "^6.4.18|^7.2.3"
     },
     "suggest": {
         "ext-simplexml": "It is needed for the checkstyle formatter"
@@ -85,22 +82,19 @@
     "scripts": {
         "website:copy-changelog": "@php -r \"copy('CHANGELOG.md', 'docs/changelog.md');\"",
         "website:copy-logo": "@php -r \"(is_dir('docs/.vuepress/public') || mkdir('docs/.vuepress/public')) && copy('art/logo_mixed.gif', 'docs/.vuepress/public/logo.gif') && copy('art/heart.svg', 'docs/.vuepress/public/heart.svg') && copy('art/heart.png', 'docs/.vuepress/public/heart.png');\"",
-        "csfixer:test": "php-cs-fixer fix -v",
+        "csfixer:test": "PHP_CS_FIXER_IGNORE_ENV=true php-cs-fixer fix -v",
         "phpstan:test": "phpstan analyse --ansi",
         "phpunit:test": "phpunit --colors=always",
-        "rector:test": "rector process --ansi",
         "insights": "bin/phpinsights analyse --ansi -v --no-interaction",
         "test": [
             "@phpstan:test",
             "@csfixer:test --dry-run",
-            "@rector:test --dry-run",
             "@phpunit:test",
             "@insights"
         ],
         "fix": [
             "@csfixer:test",
-            "@insights --fix --quiet",
-            "@rector:test"
+            "@insights --fix --quiet"
         ],
         "post-install-cmd": [
             "@website:copy-changelog",
