@@ -9,7 +9,7 @@ return $config
         'array_syntax' => ['syntax' => 'short'],
         'no_empty_statement' => true,
         'no_unneeded_control_parentheses' => true,
-        'no_unneeded_curly_braces' => true,
+        'no_unneeded_braces' => true,
         'no_unused_imports' => true,
         'ordered_imports' => true,
         'protected_to_private' => true,
