@@ -2,41 +2,17 @@
 
 declare(strict_types=1);
 
-use Rector\Core\Configuration\Option;
-use Rector\Core\ValueObject\PhpVersion;
-use Rector\DeadCode\Rector\ClassMethod\RemoveUnusedPromotedPropertyRector;
-use Rector\Set\ValueObject\SetList;
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+use Rector\Config\RectorConfig;
 
-return static function (ContainerConfigurator $containerConfigurator): void {
-    $parameters = $containerConfigurator->parameters();
-    $parameters->set(Option::PATHS, [__DIR__ . '/src', __DIR__ . '/bin']);
-    $parameters->set(Option::BOOTSTRAP_FILES, [
-        __DIR__ . '/vendor/squizlabs/php_codesniffer/autoload.php',
-    ]);
-
-    $parameters->set(Option::PHP_VERSION_FEATURES, PhpVersion::PHP_74);
-    $parameters->set(Option::AUTO_IMPORT_NAMES, true);
-
-    // skip root namespace classes, like \DateTime or \Exception [default: true]
-    $parameters->set(Option::IMPORT_SHORT_CLASSES, false);
-    // skip classes used in PHP DocBlocks, like in /** @var \Some\Class */ [default: true]
-    $parameters->set(Option::IMPORT_DOC_BLOCKS, false);
-
-    // Path to phpstan with extensions, that PHPSTan in Rector uses to determine types
-    //$parameters->set(Option::PHPSTAN_FOR_RECTOR_PATH, getcwd() . '/phpstan.neon.dist');
-
-    $containerConfigurator->import(SetList::CODE_QUALITY);
-    $containerConfigurator->import(SetList::DEAD_CODE);
-    $containerConfigurator->import(SetList::EARLY_RETURN);
-    $containerConfigurator->import(SetList::PHP_71);
-    $containerConfigurator->import(SetList::PHP_72);
-    $containerConfigurator->import(SetList::PHP_73);
-    $containerConfigurator->import(SetList::PHP_74);
-
-    $parameters->set(Option::SKIP, [
-        \Rector\EarlyReturn\Rector\If_\ChangeAndIfToEarlyReturnRector::class,
-        \Rector\CodeQuality\Rector\Identical\SimplifyBoolIdenticalTrueRector::class,
-        RemoveUnusedPromotedPropertyRector::class,
-    ]);
-};
+return RectorConfig::configure()
+    ->withPaths([
+        __DIR__.'/src',
+    ])
+    ->withPreparedSets(
+        deadCode: true,
+        codeQuality: true,
+        typeDeclarations: true,
+        privatization: true,
+        earlyReturn: true,
+        strictBooleans: true,
+    )->withPhpSets(php81: true);
